Moved non api logic to ConceptSetService file
