+++ js/modules/WebAPIProvider/ConceptSetAPI.js
@@ -1,9 +1,9 @@
 define(function (require, exports) {
 
 	var $ = require('jquery');
+	var ko = require('knockout');
 	var config = require('appConfig');
 	var authApi = require('webapi/AuthAPI');
-	var ko = require('knockout');
 
 	function getGenerationInfo(conceptSetId) {
 		var infoPromise = $.ajax({
@@ -12,8 +12,8 @@ define(function (require, exports) {
 		});
 		return infoPromise;
 	}
-		
-		function deleteConceptSet(conceptSetId) {
+
+	function deleteConceptSet(conceptSetId) {
 		var promise = $.ajax({
 			url: config.webAPIRoot + 'conceptset/' + (conceptSetId || '-1') ,
 			method: 'DELETE',
@@ -58,52 +58,12 @@ define(function (require, exports) {
 		});
 	}
 
-	function getIncludedConceptSetDrawCallback({ model, searchConceptsColumns }) {
-		return async function (settings) {
-			if (settings.aoData) {
-				const api = this.api();
-				const rows = this.api().rows({page: 'current'});
-				const data = rows.data();
-				await model.loadAndApplyAncestors(data);
-				const columnIndex = searchConceptsColumns.findIndex(v => v.data === 'ANCESTORS');
-				api.cells(null, columnIndex).invalidate();
-				rows.nodes().each((element, index) => {
-					const rowData = data[index];
-					model.contextSensitiveLinkColor(element, rowData);
-					const context = ko.contextFor(element);
-					ko.cleanNode(element);
-					ko.applyBindings(context.createChildContext(rowData), element);
-				})
-			}
-		}
-	}
-
-	function getAncestorsModalHandler({ model, ancestors, ancestorsModalIsShown }) {
-		return function(conceptId) {
-			ancestors(model.includedConcepts()
-				.find(v => v.CONCEPT_ID === conceptId)
-				.ANCESTORS
-				.map(v => ({concept: v})));
-			if (!_.isEmpty(ancestors())) {
-				ancestorsModalIsShown(true);
-			}
-		}
-	}
-	
-	function getAncestorsRenderFunction() {
-		return (s, p, d) => `<a data-bind="click: function() {$parents[2].showAncestorsModal(${d.CONCEPT_ID});}, tooltip: '${d.ANCESTORS.map(d => d.CONCEPT_NAME).join('<br>')}'" class="clickable">${d.ANCESTORS.length}</a>`;
-		
-	}
-
-	const api = {
-		getGenerationInfo: getGenerationInfo, 
+	var api = {
+		getGenerationInfo: getGenerationInfo,
 		deleteConceptSet: deleteConceptSet,
 		exists: exists,
-		saveConceptSet: saveConceptSet,
-		saveConceptSetItems: saveConceptSetItems,
-		getIncludedConceptSetDrawCallback: getIncludedConceptSetDrawCallback,
-		getAncestorsModalHandler: getAncestorsModalHandler,
-		getAncestorsRenderFunction: getAncestorsRenderFunction,
+		saveConceptSet,
+		saveConceptSetItems,
 	};
 
 	return api;
